+++ pyoxidizer/src/starlark/python_executable.rs
@@ -13,7 +13,6 @@ use {
             PythonPackageDistributionResourceValue, PythonPackageResourceValue,
             ResourceCollectionContext,
         },
-        target::{BuildContext, BuildTarget, ResolvedTarget, RunMode},
         util::{
             optional_dict_arg, optional_list_arg, required_bool_arg, required_list_arg,
             required_str_arg, ToOptional,
@@ -38,6 +37,7 @@ use {
             starlark_signature_extraction, starlark_signatures,
         },
     },
+    starlark_dialect_build_targets::{BuildContext, BuildTarget, ResolvedTarget, RunMode},
     std::{
         collections::HashMap,
         io::Write,